ContextManagement/pyproject.toml
[project]
name = "contextmanagement"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncio>=4.0.0",
    "openai-agents>=0.2.6",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
]
